Fix closure leak in BrowserCodeReader
decodeContinuously() and decodeOnce() each have a loop() function that is used
with setTimeout() to repeatedly check if a barcode is detected in the input.
loop() is wrapped in arrow functions before it is passed to setTimeout(). This
causes closure leak because the new closure contains reference to the closure
from the previous call, which in turn contains reference to closure from an
even earlier call, preventing them from being garbage-collected. Fix this by
passing loop() directly to setTimeout() instead.