Added guard for when raw bytes aren't set
UPCEANReader supplies null and 0 for rawBytes and length respectively (see src/core/oned/UPCEANReader.ts line 103). If EAN13 and UPC_A are both enabled and the barcode starts with a 0, then line 103 of the conversion here threw an exception at rawBytes.length which was caught and converted into NotFoundException higher up.